Fix broken grace window check in CanBeRenewed
https://eaflood.atlassian.net/browse/WC-512

When refactoring the grace window functionality in the `CanBeRenewed` concern we didn't spot that doing so would actually break the app. The critical we didn't take into account in our tests was that we were not setting the status to 'EXPIRED'.

By not doing so the tests would pass, but when run for real where registrations are marked as EXPIRED it then was hitting the status check and failing. This would cause `can_renew?` to return false plus again ascribe the wrong error message.

This change updates the tests to ensure the registration being checked actually represents what is expected when used in production, and fixes the broken logic in `CanBeRenewed.can_renew?`.

🎉 WIP 🎉