Add test for possible grace window loophole
When mulling on the updated logic had a feeling that we were exposing a possible hole, i.e. because we no longer rely solely on the status for checking for expired registrations with the introduction of the grace window, a REVOKED but expired registration could become eligible for renewal.

So add this test to confirm my suspicion and found this to be the case (test fails).

Follow up to this will implement the fix.