Implement logic to stop renew expired revoked regs
This change ensures that a REVOKED and expired registration that happens to be in the expiry grace window cannot be renewed.

We've moved the status check back above the expiry checking in `CanBeRenewed.can_renew?` so we don't have to mess with things too much.

So now if you are not ACTIVE or EXPIRED, you cannot renew because those are the only 2 states a registration is valid to be renewed. Hence in our case a REVOKED registration is immediately blocked.

We then tweak the expiry check back to what it was i.e. if the registration is in the grace window, we don't return immediately but instead drop allow the method to complete and return true.