Refactor can_renew? functionality to service
Rightly highlighted in the PR feedback was that the `can_renew?` method had become quite complex and therefore not easy to understand what was going on.

It was suggested I move each of the sections into helper methods like, for example

```ruby
def can_renew?(log_reason = false, error_id = :can_renew)
    if lower?
      add_validation_error(:registration_is_lower_tier, error_id) if log_reason
      return false
    end

    ...
```

Would become

```ruby
def can_renew?(log_reason = false, error_id = :can_renew)
    return false if tier_is_not_renewable?

    ...
```

What I found when I came to do this though, was that the validation error handling logic didn't make sense in the new context. So we create a new method on the concern called `tier_is_not_renewable?`, but because we are trying to simplify `can_renew?` we also take the logic of adding a validation error message to the registration's errors collection if `log_reason` is true. That is fine in the context of it being called from `can_renew?`, but as its now a method on the registration it's free to be called from anywhere.

I could have left the validation error logic in `can_renew?`, but then the new methods became single line statements offering little value in the refactor.

Hence I chose instead to go with a design where all the logic for determining whether something can be renewed is moved off to a separate class (a service object in this case). We can then control the expected context, and the API exposed.

The result is hopefully something that is easier to grok than the previous iteration, but where we don't pollute the registration with methods that actually should only be used in certain circumstances.