Ensure current day is after expiry & b4 grace window
Based on some scenarios encountered working on the waste-carriers-renewals project, which is also getting a 3 day grace window, updated the `in_expiry_grace_window?`.

It now not only checks the current day is before the last day of the grace window, but that it's also after the date the registration expires.

Also added a test to ensure we have covered the grace window being set to 0 i.e. there is no grace window.