Fix broken grace window check in CanBeRenewed (#175)
https://eaflood.atlassian.net/browse/WC-512

When refactoring the grace window functionality in the `CanBeRenewed` concern we didn't spot that doing so would actually break the app. The critical scenario we didn't take into account in our tests was that we were not setting the status to 'EXPIRED'.

By not doing so the tests would pass, but when run for real where registrations are marked as EXPIRED it then was hitting the status check and failing. This would cause `can_renew?` to return false plus again ascribe the wrong error message.

This change updates the tests to ensure the registration being checked actually represents what is expected when used in production, and fixes the broken logic in `CanBeRenewed.can_renew?`. We also

- added a test and solution for a possible loop hole spotted where with the introduction of the grace window, a REVOKED but expired registration could become eligible for renewal.
- ensure it checks the current day is before the last day of the grace window, but that it's also after the date the registration expires (blocks a loophole spotted when implementing this in waste-carriers-renewals).
- added a test to ensure we have covered the grace window being set to 0 i.e. there is no grace window.
- refactored `can_renew?` in the `CanBeRenewed` concern to a service, as it had become quite complex with all its various checks plus the handling of the validation errors