Add link to transfer a registration to different account (#176)
* Add link to transfer a registration to different account

https://eaflood.atlassian.net/browse/WC-61

NCCC users can transfer a registration (and its transient registration, if one exists) to a different account. If this account doesn't exist yet, it is created.

This feature has been built in waste-carriers-back-office, but we need to add a link to the backend. This should only be visible to agency users, not finance users.

* Generate URL to transfer a registration

When given a registration, build a URL to the page which allows transferring the registration.

This also required adding a new URL for the back office root to the application config, as we only had one specific to doing a renewal.

* Add transfer link to backend dashboard

This link is only viewable for non-finance users.