Use NCCC default for AD accountEmail instead of individual emails
https://eaflood.atlassian.net/browse/WC-482

Currently AD registrations set the accountEmail to the individual email of the agency_user. This means that these agency_users are treated as the account holder by the apps, and end up receiving email notifications about registrations they happened to register.

To stop this happening, we want to use a default shared inbox for AD registrations instead.