Install Kaminari, add default config file
We use Kaminari for pagination in our other waste carriers Rails apps. Our intention is that by only loading a limited number of registrations at a time, this will avoid running into the timeout issue. It hopefully also has the benefit of not forcing a user to scroll through hundreds or thousands of items at once.