Fix styling bug
Clicking on the pagination links caused them to move up a few pixels, shifting the rest of the link list out of alignment. This turned out to be down to a difference in the template versions between the frontend and back-office, so we need an addition to the styling to fix it.