Move passenger to production only in Gemfile
Passenger was bumped to the main group in the Gemfile to replace Webrick as the app server when running the app locally.

The intention was to more closely mirror our setup in production. However on review we have found that there have not been any times where this has helped us resolve an issue or develop a new feature. What the dev's often do though is comment it out before running locally so things like byebug will work.

So rather than hinder local development for no gain, we have chosen to stop passenger locally in our Vagrant build.