Implement location question into journey
Specifically this just deals with the main location question and not the others like Wales and Scotland.

So we have a controller, a route, locales and validation, and the other controllers that route to this page have been updated.

Also it sticks with the pattern currently implemented in the frontend. So we at least have a separate controller and view, but it does stick with using hard coded routing and same structure for the view which have since improved on.