Fix broken authentication features
PR #179 updated the frontend to redirect users to the new front office application when they attempted to sign in.

This is because the front office [PR #148](https://github.com/DEFRA/waste-carriers-front-office/pull/148) update the front office with a new user dashboard, so the one on the frontend is now redundant.

What we hadn't done is pick up on the feature tests that still expect the external user dashboard to be hosted in the project, and which now fail because of the change.

So this change removes any features that specifically cover external user authentication (as these are now covered elsewhere) and amends any that rely on a signed in user.