Remove default cucumber.yml config
Hadn't realised that this was there and that it was changing our expectations of what features were and were not being run.

Also the way it defines the tags is now deprecated. So deleting it as its no longer needed and as shown can cause confusion for users not aware that its there.