Fix and update broken cucumber features (#183)
A change we made in PR #179 meant that the external user login page is no longer used in this application and instead is hosted in the [Waste Carriers Front Office](https://github.com/DEFRA/waste-carriers-front-office).

Now any calls to it in this app redirects to the front office. What we hadn't appreciated when we made this change is just how many cucumber features rely being able to login and access the dashboard within this project.

So when we came to run the features recently a number of them were broken. On top of that we realised some were never getting called because of the default profile we have commited to the project.

So this PR is essentially a 'spring clean' of the features. We have removed those

- we can no longer support because they rely on external apps
- that were marked `@bad-test`, `@quanrantine` or `@wip` as they were never getting called anyway
- that were actually just used for testing page object functionality
- those that tested IR functionality. The last renewable IR registration was in Sept 2018, so though the code remains (its too strongly coupled to consider removing) we have removed the features as no user now can ever use this functionality now

Along with this we have removed all unused steps by using the `bundle exec cucumber -f stepdefs` to identify them, and implemented those that were undefined (if we could).

We have removed the default `cucumber.yml` file to avoid confusion and situations where we assume all scenarios are being run but aren't.

Finally we have updated cucumber to bring the project up to date.