Fix specs by re-recording VCR cassettes
It's a known issue with the VCR cassettes for companies house that edge cases like companies in voluntary arrangement won't remain in that state for long periods of time.

Recently when running the unit tests we got exactly that issue in the the previous example has now gone into administrtation.

So this change updates the specs and VCR to use a different company that is currently listed as being in Voluntary Arrangement.