Add location question to journey (#182)
Add location question to journey

https://eaflood.atlassian.net/browse/RUBY-214

As part of building the renewals service we added support for a location question. The main purpose is to

- ensure that the user is registered with the right authority
- reduce the burden on EA staff having to check for users based elsewhere who have registered with us, who they then have to confirm if this is correct.

The new registration process prior to this change did not have the question. Adding the question to renewals has helped there, but to ensure all users both internal and external benefit this change adds the question to the waste carriers frontend.

In summary

- add the main location question controller, view and unit tests
- add the pages for the other country options (Wales, Scotland and Northern Ireland)
- update connecting pages so location comes between the start page and the business type page
- fix any broken unit and acceptance tests