Remove the env var that blocks simplecov
We want simeplcov to generate test coverage information in order that this can then be sent to CodeClimate.

However this project uses an env var to control where simplecov is required when rspec runs.

We don't believe it actually adds that much time to the test run, so rather than add the env var to travis this change brings the project inline with the rest and has simplecov always being used.