Remove extraneous simeplcov config and only run rspec
The unit tests are now all passing in Travis but it still records an error. This is because the current config ropes cucumber in when `bundle exec rake` is called. This is what travis-ci defaults to.

So we have tweaked the Travis config to resolve that.

Also there was additional simplecov config and setup going on that again we have not used in our other projects.

So this change also brings the setup for simplecov inline with those other projects.