Add Continuous Integration support (#195)
With the repo being cleansed and the made public we can now hook it up to our usual suite of CI tools.

This changes adds the additional config necessary and most importantly badges to the README 😁

To support ensuring we get test coverage this change also includes sorting out how simplecove is set up in this project to bring it inline with our others.

Another thing to note is we have to specify that travis runs `bundle exec rspec`. This is because the current cucumber rake task adds itself to the default rake task which is what Travis-Ci would normally call.

We can't support running the cucumber tests in Travis at this time, and it will take some additional work to sort out the cucumber rake task hence for now we just update the Travis config.