Update Money-rails to latest
We need to update i18n to resolve a security issue. However when we do so it causes an error in [Money-rails](https://github.com/RubyMoney/money-rails).

Upgrading money-rails resolves the error hence this change.

One issue we faced is that Money-rails is essentially a rails wrapper for the [money](https://github.com/RubyMoney/money) gem, and it has changed its default beahviour.

It now requires you to specify that you wish to use I18n localization as part of the gem config as per https://github.com/RubyMoney/money#deprecation