Fix how mongoid is specified in Gemfile
In PR #204 (https://github.com/DEFRA/waste-carriers-frontend/pull/204) Dependabot attempted to update mongoid to 5.4.0. However on checking with our other WCR projects the version they are using is Mongoid 5.2.1.

The one difference between the projects was the difference in how the versioning was specified in the Gemfile. We want to keep the version of mongoid consistent across the projects because we are so dependant on it hence this change to bring the frontend in line with the Front office and Back office projects.