Update rails version and Gemfile declaration (#213)
We need to update the rails version to resolve a critical security issue in the previous version of ActionView (see [CVE-2019-5418](https://groups.google.com/forum/#!topic/rubyonrails-security/pFRKI96Sm8Q))

This also updates the way we specify the version in the Gemfile. If we specified it as we have done now this would have been picked up by Dependabot, but as it is we'd just get a warning and then manually have to intervene to update it.

So making this change brings the project inline with the other apps for this service.