Add email interception and retrieval
In order to improve the flow of the end-to-end tests, we want to add behaviour to the application that will support the storage of the last email that was sent. There will also be a route that returns the latest mail object as JSON.

The feature will be enabled through the use of an ENV var so we can ensure the endpoint is only accessible when the feature is explicitly enabled.