Fix test & only intercept when necessary
**Only intercept when necessary**

This was the original intention but after many attempts realised that the initializers are called before any kind of mocking/stubbing goes on in the RSpec test. So we would never be able to control whether the interceptor is registered within our tests.

So next we looked at ignoring the interceptor and using either RSpec filters or hooks to control the registering and unregistering of the interceptor. However we hit an issue there as well. Registering the interceptor was fine, but unregistering was only merged into ActionMailer in May 2018 (https://github.com/rails/rails/pull/32207).

So we'd be left with having to also introduce a monkey patch to ensure the tests were clean and the interceptor was removed. 😩

This seemed overkill for what we wanted to achieve. Hence we hit on a final hack, which is to always register the interceptor in the initializer if we are running in "test".

**Fixing the test**

Realised when working through trying to control whether the interceptor is registered or not that the previous version of the test wasn't actually testing its intent. It is based on the fact that the last email cache feature is enabled, and as such when we check the response we expect to see some email content.

However when the interceptor was not being registered `LastEmailCache` was just returning its default error message. As the test was just comparing the response to `LastEmailCache.last_email_json` they were always going to match whether the interceptor was registered or not.

So this also tweaks the test such that it matches the intent of checking that the response contains content from an actual sent email.