Refactor the rake test email task
With the changes made in [PR 227](https://github.com/DEFRA/waste-carriers-frontend/pull/227) and [PR 228](https://github.com/DEFRA/waste-carriers-frontend/pull/228) we no longer need to use a mail in the `RegistrationMailer` class for testing if email is setup and working.

We now have a dedicated `TestMailer` with associated templates that can be used for this.

So this change is just a bit of housekeeping (akin to rearranging deck chairs on the titantic!)