Use CGI.escape for Worldpay urls
https://eaflood.atlassian.net/browse/RUBY-729

Whilst working with Worlpay to move to their new hosted pages service, it was spotted that the urls the frontend tells Wordpay to redirect users back to when payment is complete were not being escpaed properly.

Worldpay advise that they are escaped to ensure there are no issues.

The [waste-carriers-engine](https://github.com/DEFRA/waste-carriers-engine) uses `CGI.escape` instead of `URI::encode` and we found in testing that it then matches what worldpay expect.