Remove renewal extension message
We need to update the path used to redirect users to the front office (and back office for internal users) when starting a renewal. As part of that we came across this message which was used to explain to users who tried to renew prior to September 2018 why it was not available.

Waste Carrier Renewals have been live and working since September 2018. So we no longer need the message hence this change removes it. It should also make focusing one what exactly we do need to change a little easier.