Change redirect path for renewals
We had to implement a change to how urls, and specifically any url that points to a page which handles transient registrations are formed.

https://github.com/DEFRA/waste-carriers-engine/pull/579

Essentially we needed to switch to using a token rather than the registration identifier. That specific change does not affect the frontend, but as part of the same PR we also fixed an issue with validation errors, where generating one would break back links.

This turned out to just be because we were putting the ID for a 'resource' at the end of the url, rather than before the action we wanted to carry out against it. This seems to break a rails convention hence generated back links stopped working in certain cases.

That fix does impact this project, because it now needs to ensure that the path it redirects users to when renewing matches what the new front and back office apps expect.

This change covers everything needed to update the frontend to redirect users correctly.