Change redirect path for renewals (#279)
We had to implement a change to how URLs, and specifically any URL that points to a page which handles transient registrations are formed.

https://github.com/DEFRA/waste-carriers-engine/pull/579

Essentially we needed to switch to using a token rather than the registration identifier. That specific change does not affect the frontend, but as part of the same PR, we also fixed an issue with validation errors, where generating one would break backlinks.

This turned out to just be because we were putting the ID for a 'resource' at the end of the URL, rather than before the action we wanted to carry out against it. This seems to break a rails convention hence generated back links stopped working in certain cases.

That fix does impact this project because it now needs to ensure that the path it redirects users to when renewing matches what the new front and back-office apps expect.

This change covers everything needed to update the frontend to redirect users correctly.

N.B. This also removes the check whether to show the renewals holding page

PR #278 removed the renewals holding page. We should have probably removed the check whether to show the holding page as well, but forgot we had one!