Resolve issue in base_office_url method
When we made the changes in https://github.com/DEFRA/waste-carriers-frontend/pull/279 we mistakenly thought the `/bo` and `/fo` paths were only needed in production, and were not to be used when running locally.

This is incorrect and they should be used in all environments.

So this removes the check for production in `base_office_url()`.