Fix back_office_renewals_url and refactor
This fix sorts the link we generate to start a back office renewal. It also take the opportunity to refactor how we generate the link to match what we have done elsewhere.

So this drops it out of the concern (what plonker put it there!) and adds it to the view helper instead.