Remove test email rake task
Now we have [defra-ruby-email](https://github.com/DEFRA/defra-ruby-email) implemented we also get a test email rake task included with it as well.

So there is no need to keep the rake task we have in this project, or the `TestMailer` anymore (some previous specs and the rake task were the only things depending on it).