Redirect payment details to back office (#303)
https://eaflood.atlassian.net/browse/RUBY-936

We have completely replaced all financial functionality in the service with new functions in the back-office. So going forward we want users to be directed there to use its functions, and not using the old ones in this app.

To help with that, when users click the 'Payment details' link we want them to be directed to the back office. Hence this change.