Redirect view certificate to back office (#304)
https://eaflood.atlassian.net/browse/RUBY-937

We have replaced the view certifcate feature in the service with a new version in the back-office. So going forward we want users to be directed there to use it, instead of the old version in this app.

To help with that, when users click the 'View certifcate' link we want them to be directed to the back office. Hence this change.