Switch to SonarCloud from CodeClimate (#306)
https://sonarcloud.io/organizations/defra/projects

We currently use [Code Climate](https://codeclimate.com/) for our code quality checks and test coverage analysis. However a number of Defra projects have used an internal SonarQube instance, and some public ones have been using SonarCloud (the cloud SASS version of SonarQube).

It looks like we're about to formally adopt SonarCloud and all projects will coalesce onto. So to keep on top of things and continue to be an example to others, we are working through our repos and switching them to SonarCloud.

** Notes

The figuring out of how to do this (because it wasn't straight forward!) was done in

- https://github.com/DEFRA/waste-exemptions-front-office/pull/317
- https://github.com/DEFRA/waste-exemptions-front-office/pull/318

*** Rubocop not part of CI

In all our other projects we run Rubocop as part of the CI. Turns out this project was not set up to run Rubocop as part of CI, most likely because it would fail abysmally! So we have had to tweak the SonarCloud config to not include Rubocop or its output.