Add feature toggles to frontend
This doesn't add defra-ruby-features because we do not intend to support managing the features from here. But it does add the same `FeatureToggle` class used in the engine.

It has the power to read from a config file or the database, with the db taking precedence. With this in place we can use it as part of the logic in the controller to determine whether new registration requests should be redirected to the front-office.