Toggle redirect to FO new reg. using feat. toggle (#329)
https://eaflood.atlassian.net/browse/RUBY-1118

In preparation for launching the 'new' registration journey in the [waste-carriers-front-office](https://github.com/DEFRA/waste-carriers-front-office) we added an automatic redirect to it here.

You can override it with a param, and this was going to be our fall back should an issue be found that means we need to quickly revert the change. We would still have been left with the issue of external users who would not know about the param. To resolve that we'd either be forcing everyone to call NCCC and chasing down an emergency RfC.

However, we have now updated our feature toggle support in the new apps to be controllable from the app themselves rather than just env vars or .yml files. We decided wouldn't it be great if we could also control that redirect based on the current setting of our `new_registration` toggle. Then there would be no need for back-door params or emergency RfC's and panic in NCCC.

This change is about getting this project to be able to read that feature toggle and control its re-direct based on it.